Fix wrong deletion ordering for Users
We were deleting the users with `foreground` cascading. Foreground
cascading ensures that an owner is deleted LAST.

However, we want the owner `User` object to be deleted FIRST. So the
`foreground` deletion cascade is actually the wrong way around.

By using the default `background` cascade, it will delete the owner
first and only after it's gone, the owned objects.

Refs:
- https://kubernetes.io/docs/tasks/administer-cluster/use-cascading-deletion/#use-foreground-cascading-deletion
- https://kubernetes.io/docs/concepts/overview/working-with-objects/owners-dependents/